4.1. Продолжение знакомства с SQL

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -61,7 +61,7 @@ public class StudentController {
     }
 
     @GetMapping("/average-age")
-    public ResponseEntity<Integer> studentsAverageAge(){
+    public ResponseEntity<Double> studentsAverageAge(){
         return ResponseEntity.ok(this.studentService.studentsAverageAge());
     }
